refactor(modules/**/*spec.ts): refactor spec files to be used by karma and move to test dir
