+++ test/preboot/server/presets_spec.js
@@ -1,70 +1,63 @@
-/// <reference path="../../typings/tsd.d.ts"/>
-
-import presetFns from '../../src/server/presets';
-
+var presets_1 = require('../../../dist/preboot/src/server/presets');
 /**
  * These tests are pretty basic, but just have something in
  * place that we can expand in the future
  */
 describe('presets', function () {
-
     describe('keyPress()', function () {
         it('should add listen selector', function () {
-            let opts = { listen: [] };
-            let expected = {
+            var opts = { listen: [] };
+            var expected = {
                 listen: [{
-                    name: 'selectors',
-                    eventsBySelector: {
-                        'input[type="text"],textarea': ['keypress', 'keyup', 'keydown']
-                    }
-                }]
+                        name: 'selectors',
+                        eventsBySelector: {
+                            'input[type="text"],textarea': ['keypress', 'keyup', 'keydown']
+                        }
+                    }]
             };
-            presetFns.keyPress(opts);
+            presets_1.default.keyPress(opts);
             expect(opts).toEqual(expected);
         });
     });
-
     describe('focus()', function () {
         it('should add listen selector', function () {
-            let opts = { listen: [] };
-            let expected = {
+            var opts = { listen: [] };
+            var expected = {
                 listen: [{
-                    name: 'selectors',
-                    eventsBySelector: {
-                        'input[type="text"],textarea':   ['focusin', 'focusout']
-                    },
-                    trackFocus: true,
-                    doNotReplay: true
-                }]
+                        name: 'selectors',
+                        eventsBySelector: {
+                            'input[type="text"],textarea': ['focusin', 'focusout']
+                        },
+                        trackFocus: true,
+                        doNotReplay: true
+                    }]
             };
-            presetFns.focus(opts);
+            presets_1.default.focus(opts);
             expect(opts).toEqual(expected);
         });
     });
-
     describe('buttonPress()', function () {
         it('should add listen selector', function () {
-            let opts = { listen: [], freeze: { name: 'spinner', eventName: 'yoyo' } };
-            let expected = {
+            var opts = { listen: [], freeze: { name: 'spinner', eventName: 'yoyo' } };
+            var expected = {
                 listen: [{
-                    name: 'selectors',
-                    preventDefault: true,
-                    eventsBySelector: {
-                        'input[type="submit"],button': ['click']
-                    },
-                    dispatchEvent: opts.freeze.eventName
-                }],
+                        name: 'selectors',
+                        preventDefault: true,
+                        eventsBySelector: {
+                            'input[type="submit"],button': ['click']
+                        },
+                        dispatchEvent: opts.freeze.eventName
+                    }],
                 freeze: { name: 'spinner', eventName: 'yoyo' }
             };
-            presetFns.buttonPress(opts);
+            presets_1.default.buttonPress(opts);
             expect(opts).toEqual(expected);
         });
     });
-
     describe('pauseOnTyping()', function () {
         it('should add listen selector', function () {
-            let opts = { listen: [], pauseEvent: 'foo', resumeEvent: 'choo' };
-            let expected = {
+            var opts = { listen: [], pauseEvent: 'foo', resumeEvent: 'choo' };
+            var expected = {
                 listen: [
                     {
                         name: 'selectors',
@@ -88,9 +81,8 @@ describe('presets', function () {
                 pauseEvent: opts.pauseEvent,
                 resumeEvent: opts.resumeEvent
             };
-            presetFns.pauseOnTyping(opts);
+            presets_1.default.pauseOnTyping(opts);
             expect(opts).toEqual(expected);
         });
     });
-
 });
